moved tuples to separate module
