+++ kotlin-spark-api/3.2/pom_2.12.xml
@@ -27,6 +27,10 @@
       <groupId>org.jetbrains.kotlinx.spark</groupId>
       <artifactId>core-3.2_${scala.compat.version}</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.jetbrains.kotlinx.spark</groupId>
+      <artifactId>scala-tuples-in-kotlin</artifactId>
+    </dependency>
 
 
     <!-- Provided dependencies -->
@@ -75,6 +79,7 @@
     <testSourceDirectory>src/test/kotlin</testSourceDirectory>
     <directory>target/${scala.compat.version}</directory>
     <plugins>
+
       <plugin>
         <groupId>org.jetbrains.kotlin</groupId>
         <artifactId>kotlin-maven-plugin</artifactId>
@@ -93,10 +98,12 @@
           </execution>
         </executions>
       </plugin>
+
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
       </plugin>
+
       <plugin>
         <groupId>org.jetbrains.dokka</groupId>
         <artifactId>dokka-maven-plugin</artifactId>
@@ -121,6 +128,7 @@
           </execution>
         </executions>
       </plugin>
+
       <plugin>
         <groupId>io.qameta.allure</groupId>
         <artifactId>allure-maven</artifactId>
@@ -128,10 +136,12 @@
           <resultsDirectory>${project.basedir}/allure-results/${scala.compat.version}</resultsDirectory>
         </configuration>
       </plugin>
+
       <plugin>
         <groupId>org.jacoco</groupId>
         <artifactId>jacoco-maven-plugin</artifactId>
       </plugin>
+
     </plugins>
   </build>
 </project>
